requirements.txt
Flask==2.0.0
Flask-Session==0.4.0
joblib==1.1.0
jsonschema==4.1.2
gunicorn==20.1.0
Jinja2==3.0.2
mysql-connector==2.2.9
requests==2.26.0
Werkzeug==2.0.2
gunicorn==20.1.0
pandas==1.3.4
openpyxl==3.0.9
